Feat: make it compile and run tests git push
